finished sync

+++ AgroMap/AgroMap/AgroMap/Database/EventDAO.cs
@@ -10,62 +10,68 @@ using Xamarin.Forms;
 
 namespace AgroMap.Database
 {
-    class EventDAO
+    public static class EventDAO
     {
         private static string Table = "Event";
-        readonly SQLiteAsyncConnection db;
-
-        public EventDAO()
-        {
-            try
-            {
-                if (db == null)
-                {
-                    string dbpath = DependencyService.Get<IFileHelper>().GetLocalFilePath("agromap_database.db3");
-                    db = new SQLiteAsyncConnection(dbpath);
-                    db.CreateTableAsync<Event>().Wait();
-                }
-            }
-            catch (Exception err)
-            {
-
-                Debug.WriteLine("AGROMAP|EventDAO(): " + err.Message);
-            }
-                
-        }
         
-        public async  Task<List<Event>> GetItens()
+        public static async Task<List<Event>> GetAll()
         {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if(db==null)
+                return null;
+            await CheckTable();
             return await db.Table<Event>().ToListAsync();
         }
 
-        public async Task<List<Event>> GetItensBy(string param, string value, Boolean isInt)
+        public static async Task<List<Event>> GetEventsByInspection(int inspection)
         {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if (db == null)
+                return null;
+            await CheckTable();
             try
             {
-                int __value = 0;
-                if (isInt)
-                {
-
-                    __value = Convert.ToInt32(value);
-                }
-                string sql = String.Format("SELECT * FROM {0} WHERE {1} = {2}", Table, param, __value);
-                Debug.WriteLine("AGROMAP|EventDAO.cs|GetItensBy: SQL: " + sql);
+                string sql = String.Format("SELECT * FROM {0} WHERE inspection = {1}", Table, inspection);
 
                 return await db.QueryAsync<Event>(sql);
             }
             catch (Exception err)
             {
-                Debug.WriteLine("AGROMAP|EventDAO.cs|GetItensBy: " + err.Message);
+                Debug.WriteLine("AGROMAP|EventDAO.cs|GetEventsByInspection: " + err.Message);
                 return null;
             }
         }
 
-        public async Task<Event> GetByID(int id)
+        public static async Task<List<Event>> GetUnsyncedByInspection(int inspection)
         {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if (db == null)
+                return null;
+            await CheckTable();
             try
             {
-                return await db.Table<Event>().Where(i => i.Id == id).FirstOrDefaultAsync();
+
+                string sql = String.Format("SELECT * FROM {0} WHERE inspection = {1} AND synced = 0", Table, inspection);
+
+                return await db.QueryAsync<Event>(sql);
+            }
+            catch (Exception err)
+            {
+                Debug.WriteLine("AGROMAP|EventDAO.cs|GetUnsyncedByInspection: " + err.Message);
+                return null;
+            }
+        }
+
+        public static async Task<Event> GetByID(int id)
+        {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if (db == null)
+                return null;
+            await CheckTable();
+
+            try
+            {
+                return await db.Table<Event>().Where(i => i.id == id).FirstOrDefaultAsync();
             }
             catch (Exception err)
             {
@@ -74,8 +80,12 @@ namespace AgroMap.Database
             }
         }
 
-        public async Task<Boolean> Update(Event item)
+        public static async Task<Boolean> Update(Event item)
         {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if (db == null)
+                return false;
+            await CheckTable();
             try
             {
                 await db.UpdateAsync(item);
@@ -88,8 +98,43 @@ namespace AgroMap.Database
             }
         }
 
-        public async Task<Boolean> Insert(Event item)
+        public static async Task<Boolean> SetSynced(List<Event> events)
         {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if (db == null)
+                return false;
+            await CheckTable();
+            try
+            {
+                foreach(Event e in events)
+                {
+                    e.synced = 1;
+                    await db.UpdateAsync(e);
+                }
+                return true;
+            }
+            catch (Exception err)
+            {
+                Debug.WriteLine("AGROMAP|EventDAO.cs|SetSynced: " + err.Message);
+                return false;
+            }
+        }
+
+        public static async Task<Boolean> Create(Event item)
+        {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if (db == null)
+                return false;
+            await CheckTable();
+            try
+            {
+
+            }
+            catch (Exception)
+            {
+
+                throw;
+            }
             try
             {
                 await db.InsertAsync(item);
@@ -98,13 +143,17 @@ namespace AgroMap.Database
             }
             catch (Exception err)
             {
-                Debug.WriteLine("AGROMAP|EventDAO.cs|Insert: " + err.Message);
+                Debug.WriteLine("AGROMAP|EventDAO.cs|Create: " + err.Message);
                 return false;
             }
         }
 
-        public async Task<Boolean> Delete(Event item)
+        public static async Task<Boolean> Delete(Event item)
         {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if (db == null)
+                return false;
+            await CheckTable();
             try
             {
                 await db.DeleteAsync(item);
@@ -116,5 +165,52 @@ namespace AgroMap.Database
                 return false;
             }
         }
+
+        public static async Task<Boolean> SaveList(List<Event> events, int inspection)
+        {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if (db == null)
+                return false;
+            await CheckTable();
+            try
+            {
+                //Limpa tabela e armazena apenas o que foi sincronizado com o servidor
+                string sql = String.Format("DELETE FROM {0} WHERE inspection = {1}", Table, inspection);
+                await db.QueryAsync<Event>(sql);
+
+                // Se lista vazia, retorna
+                if (events == null || events.Count == 0)
+                    return true;
+
+                foreach (Event e in events)
+                {
+                    e.synced = 1;
+                    await Create(e);
+                }
+                return true;
+            }
+            catch (Exception err)
+            {
+                Debug.WriteLine("AGROMAP|EventDAO.cs|SaveInLocalStorage: " + err.Message);
+                return false;
+            }
+        }
+
+        private static async Task<Boolean> CheckTable()
+        {
+            SQLiteAsyncConnection db = Database.GetConn();
+            if (db == null)
+                return false;
+            try
+            {
+                await db.Table<Event>().Where(i => i.id == 0).FirstOrDefaultAsync();
+            }
+            catch
+            {
+                db.CreateTableAsync<Event>().Wait();
+            }
+            return true;
+        }
+        
     }
 }
